drm/rockchip: demote warning about AFBC format to debug message
This driver supports YUV pixel formats, and AFBC modifiers, but not at
the same time. When computing a plane's IN_FORMATS property
the rockchip_mod_supported() helper function will be called with such
combinations. vop_convert_afbc_format() is expected to return -EINVAL
in this case, it's not a bug.

Remove the comment about this case being unreachable, and demote the
warning to a debug message. Avoids showing a needless warning in dmesg:

    Dec 31 15:49:25 rockpro64 kernel: [drm] unsupported AFBC format[3231564e]

Signed-off-by: Dan Callaghan <djc@djc.id.au>